// lib/frontend/keyboard-shortcuts/src/type.ts
/*
 * Copyright (c) Enalean, 2020-Present. All Rights Reserved.
 *
 * This file is a part of Tuleap.
 *
 * Tuleap is free software; you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation; either version 2 of the License, or
 * (at your option) any later version.
 *
 * Tuleap is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with Tuleap. If not, see <http://www.gnu.org/licenses/>.
 */

export interface Shortcut {
    readonly keyboard_inputs: string;
    readonly displayed_inputs?: string;
    readonly description: string;
    readonly handle: null | ((event: KeyboardEvent) => void | ShortcutHandleOptions);
}

export interface ShortcutHandleOptions {
    readonly preventDefault: boolean;
}

export interface ShortcutsGroup {
    readonly title: string;
    readonly details?: string;
    readonly shortcuts: Array<Shortcut>;
}

export type Scope = "GLOBAL" | "PLUGIN";
export const GLOBAL_SCOPE: Scope = "GLOBAL";
export const PLUGIN_SCOPE: Scope = "PLUGIN";